fix: IncrementalCollectionOp on soft-deletable entities

+++ packages/orm/src/index.ts
@@ -140,6 +140,8 @@ export function setOpts<T extends Entity>(
           const allowDelete = !field.otherMetadata().fields["delete"];
           const allowRemove = !field.otherMetadata().fields["remove"];
 
+          const isSoftDeletable = !!field.otherMetadata().allFields["deletedAt"];
+
           // We're replacing the old `delete: true` / `remove: true` behavior with `op` (i.e. operation).
           // When passed in, all values must have it, and we kick into incremental mode, i.e. we
           // individually add/remove/delete entities.
@@ -154,7 +156,12 @@ export function setOpts<T extends Entity>(
             }
             values.forEach((v) => {
               if (v.op === "delete") {
-                entity.em.delete(v);
+                // We need to check if this is a soft-deletable entity, and if so, we will soft-delete it.
+                if (isSoftDeletable) {
+                  v.set({ deletedAt: new Date() });
+                } else {
+                  entity.em.delete(v);
+                }
               } else if (v.op === "remove") {
                 (current as any).remove(v);
               } else if (v.op === "include") {
